# argus/docs/swagger.yaml
basePath: /
definitions:
...
  argus_app_internal_model.User:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      deptId:
        type: integer
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      remark:
        type: string
      status:
        description: 1 启用 / 0 禁用；类型由 gorm 按驱动映射（决策 18）
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  argus_app_internal_pkg_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  argus_app_internal_service.ChangeCurrentPasswordInput:
    properties:
      newPassword:
        maxLength: 32
        minLength: 6
        type: string
      oldPassword:
        minLength: 1
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  argus_app_internal_service.CurrentProfileDTO:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      phone:
        type: string
      remark:
        type: string
      username:
        type: string
    type: object
...
  argus_app_internal_service.SaveUserInput:
    properties:
      deptId:
        type: integer
      email:
        type: string
      nickname:
        maxLength: 64
        type: string
      password:
        description: 编辑时可空，新增时需处理默认密码
        maxLength: 32
        minLength: 6
        type: string
      phone:
        maxLength: 32
        type: string
      remark:
        maxLength: 255
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      username:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - username
    type: object
  argus_app_internal_service.UpdateCurrentProfileInput:
    properties:
      avatar:
        maxLength: 255
        type: string
      email:
        type: string
      nickname:
        maxLength: 64
        type: string
      phone:
        maxLength: 32
        type: string
      remark:
        maxLength: 255
        type: string
    type: object
...
  argus_app_internal_service.UserPageItem:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      deptId:
        type: integer
      deptName:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      remark:
        type: string
      status:
        description: 1 启用 / 0 禁用；类型由 gorm 按驱动映射（决策 18）
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  argus_app_internal_service.UserPageResult:
    properties:
      items:
        items:
          $ref: '#/definitions/argus_app_internal_service.UserPageItem'
        type: array
      total:
        type: integer
    type: object
...
  internal_api.AssignRolesInput:
    properties:
      roleIds:
        items:
          type: integer
        type: array
    type: object
...
  internal_api.BatchDeleteUserInput:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
  internal_api.BatchUpdateUserStatusInput:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
      status:
        type: integer
    required:
    - ids
    - status
    type: object
...
  internal_api.NilResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_api.RoleIDsResponse:
    properties:
      code:
        type: integer
      data:
        items:
          type: integer
        type: array
      message:
        type: string
    type: object
...
  internal_api.UpdateStatusInput:
    properties:
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  internal_api.UserPageResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_service.UserPageResult'
      message:
        type: string
    type: object
  internal_api.UserResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/argus_app_internal_model.User'
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: argus 后端 API 接口文档
  title: argus API
  version: "1.0"
paths:
...
  /api/user:
    post:
      consumes:
      - application/json
      description: 创建系统用户账号，默认初始密码为 password123
      parameters:
      - description: 创建用户参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功的用户信息
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 创建新用户
      tags:
      - 用户模块
  /api/user/{id}:
    delete:
      description: 软删除指定 ID 的用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 删除指定用户
      tags:
      - 用户模块
    put:
      consumes:
      - application/json
      description: 根据 ID 修改指定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.SaveUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            $ref: '#/definitions/internal_api.UserResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户模块
  /api/user/{id}/reset-password:
    put:
      description: 将指定用户的密码重置为系统默认密码 (password123)
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 重置用户密码
      tags:
      - 用户模块
  /api/user/{id}/roles:
    get:
      description: 查询指定用户当前分配的角色 ID 列表
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 角色ID数组
          schema:
            $ref: '#/definitions/internal_api.RoleIDsResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取用户绑定的角色ID列表
      tags:
      - 用户模块
    put:
      consumes:
      - application/json
      description: 覆盖更新指定用户的角色分配关系
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色分配参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.AssignRolesInput'
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 为用户分配角色
      tags:
      - 用户模块
  /api/user/{id}/status:
    put:
      consumes:
      - application/json
      description: 启用或禁用指定用户账号
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态修改参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 修改用户状态
      tags:
      - 用户模块
  /api/user/batch:
    delete:
      consumes:
      - application/json
      description: 批量软删除指定 ID 列表的用户
      parameters:
      - description: 批量删除参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.BatchDeleteUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 批量删除用户
      tags:
      - 用户模块
  /api/user/batch-status:
    put:
      consumes:
      - application/json
      description: 批量启用或禁用指定用户
      parameters:
      - description: 批量状态修改参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.BatchUpdateUserStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 批量更新用户状态
      tags:
      - 用户模块
...
  /api/user/page:
    get:
      description: 分页查询用户数据，支持用户名、昵称、状态、部门筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 状态 (0:禁用, 1:启用)
        in: query
        name: status
        type: integer
      - description: 部门ID
        in: query
        name: deptId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户分页数据
          schema:
            $ref: '#/definitions/internal_api.UserPageResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 分页获取用户列表
      tags:
      - 用户模块
  /api/user/profile:
    get:
      description: 获取当前认证用户的基本资料（包含昵称、邮箱、手机号等）
      produces:
      - application/json
      responses:
        "200":
          description: 个人资料
          schema:
            $ref: '#/definitions/argus_app_internal_service.CurrentProfileDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 获取当前登录用户个人资料
      tags:
      - 用户模块
    put:
      consumes:
      - application/json
      description: 修改当前认证用户的昵称、邮箱、手机号、头像和个人简介
      parameters:
      - description: 修改资料参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.UpdateCurrentProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的个人资料
          schema:
            $ref: '#/definitions/argus_app_internal_service.CurrentProfileDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 修改当前登录用户个人资料
      tags:
      - 用户模块
  /api/user/profile/password:
    put:
      consumes:
      - application/json
      description: 校验旧密码并更新为新密码，同时吊销当前用户的所有 Refresh Token
      parameters:
      - description: 修改密码参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/argus_app_internal_service.ChangeCurrentPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/internal_api.NilResponse'
        "400":
          description: 参数错误或旧密码错误
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/argus_app_internal_pkg_response.Result'
      security:
      - BearerAuth: []
      summary: 修改当前登录用户密码
      tags:
      - 用户模块
securityDefinitions:
  BearerAuth:
    description: 在 Header 中传入 `Bearer <token>` 进行身份认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"